requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
python-binance==1.0.15
numpy
matplotlib
cryptography
filelock
pandas
pyarrow
apscheduler
Flask
redis
